requirements.txt
opencv-python
ultralytics
wxPython
Pillow
logging
numpy
pyinstaller
